Restructure
Restructure directory as a "real-world" application structure.
